Rework error handling and protocol update
This pull request deals with two issues that are dependent on each
other.

Error handling is reworked so it no longer uses gRPC metadata but
in-built gRPC error handling support via exceptions. gRPC submodule was
updated to reflect this change.

Proto protocol is updated. Appliance and Image sctructures was changed
and changes are described in
the-cloudkeeper-project/cloudkeeper-proto#3.
Also method UpdateAppliance was added and this change is described in
the-cloudkeeper-project/cloudkeeper-proto#4

gRPC Appliance and Image attributes will be now available in form of
OpenNebula tags for both Image and Template resources. Written values
(especially for Image) represents information about converted image file
NOT information from AppDB's image list.

Close #40