Implement single VO deployment model
As per new version (2.0) deployment model is changed to single VO
per instance.

There is no longer any concept of OpenNebula group
management in Cloudkeeper-ONE. That means no chnaging of groups on
resources will be done. This allows loose mapping of VOs and OpenNebula
groups (groups and VO doesn't have to have the same name anymore). What
group will the appliance be registered under is decided only by scoping
of the token used to access OpenNebula.

Concept of a public mode was discared as well since mapping a VO to
publicly available group can be done the same way as mapping to any
other group.